+++ unicore_gitmodels/models.py
@@ -1,6 +1,8 @@
 import re
 import unicodedata
 from gitmodel import fields, models
+from unicore_gitmodels.fields import ListField
+
 
 RE_NUMERICAL_SUFFIX = re.compile(r'^[\w-]*-(\d+)+$')
 
@@ -46,6 +48,7 @@ class GitCategoryModel(SlugifyMixin, FilterMixin, models.GitModel):
     language = fields.CharField(required=False)
     featured_in_navbar = fields.BooleanField(default=False)
     source = fields.RelatedField('GitCategoryModel', required=False)
+    position = fields.IntegerField(required=False, default=0)
 
     def __unicode__(self):
         return self.title
@@ -81,7 +84,14 @@ class GitCategoryModel(SlugifyMixin, FilterMixin, models.GitModel):
             'language': self.language,
             'featured_in_navbar': self.featured_in_navbar,
             'source': source,
+            'position': self.position,
         }
+
+    @classmethod
+    def all(cls):
+        items = list(super(GitCategoryModel, cls).all())
+        sorted_items = sorted(items, key=lambda cat: cat.position)
+        return models.ModelSet((c for c in sorted_items))
 
 
 class GitPageModel(SlugifyMixin, FilterMixin, models.GitModel):
@@ -98,6 +108,7 @@ class GitPageModel(SlugifyMixin, FilterMixin, models.GitModel):
     featured_in_category = fields.BooleanField(default=False)
     language = fields.CharField(required=False)
     source = fields.RelatedField('GitPageModel', required=False)
+    linked_pages = ListField(default=[], required=False)
 
     def __unicode__(self):
         return self.title
@@ -127,4 +138,10 @@ class GitPageModel(SlugifyMixin, FilterMixin, models.GitModel):
             'language': self.language,
             'featured': self.featured,
             'featured_in_category': self.featured_in_category,
+            'linked_pages': self.linked_pages,
         }
+
+    def get_linked_pages(self):
+        if self.linked_pages is None:
+            return []
+        return [self.__class__.get(uuid) for uuid in self.linked_pages]